Extracted asterisk configs parsing utilities to helpers.asterisk submodule
