+++ wazo_confgend/helpers/asterisk.py
@@ -3,6 +3,11 @@
 
 
 import logging
+import configparser
+import collections
+import itertools
+
+
 from wazo_confgend.generators.util import AsteriskFileWriter
 
 logger = logging.getLogger(__name__)
@@ -44,3 +49,46 @@ class AsteriskFileGenerator:
             writer.write_section(required_section)
             writer._write_section_separator()
             sections.add(required_section)
+
+
+class CustomConfigParserStorage(collections.UserDict):
+    """configparser storage class customized for asterisk format parsing"""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.data = collections.OrderedDict(self.data)
+        self.counter = itertools.count(start=1)
+
+    def __setitem__(self, k, v):
+        if isinstance(v, list):
+            # if list, we assume this is a new option declaration
+            # configparser creates option values as list to deal with multiline values
+            self.data[f"{next(self.counter)}:{k}"] = v
+        else:
+            # this is either a SectionProxy object,
+            # or an option value transformed from list form to string(so not a new option)
+            self.data[k] = v
+
+
+class AsteriskConfigParser(configparser.RawConfigParser):
+    def items(self, *args, **kwargs):
+        # Coupled with above CustomConfigParserStorage class as dict_type,
+        # we trick configparser into tracking each duplicate option declarations by storing them under different keys
+        # This reverts the trick in order to expose duplicate options under their real name
+        return [
+            (option.split(":", maxsplit=1)[-1], value)
+            for option, value in super().items(*args, **kwargs)
+        ]
+
+
+def asterisk_parser(food=None) -> AsteriskConfigParser:
+    parser = AsteriskConfigParser(
+        dict_type=CustomConfigParserStorage,
+        strict=False,
+        interpolation=None,
+        empty_lines_in_values=False,
+        inline_comment_prefixes=[";"],
+    )
+    if food:
+        parser.read_file(food)
+    return parser
